Working frontend asset pipeline.
Builds ES6+JSX format script into a single bundle.js that
runs on ES5 browsers. Currently receives events from a local
script on the page rather than a Websocket, but the Websocket
code is there and basically works.

Start of an AppCache stuff for offline support is there, but needs
to update conditionally when a resource changes (probably by including
every cached asset's shasum in a comment)

Still working out what to do with Akka