NEWSTART-19 Tiles a la Redux - idea
This commit attempts to reconcile the idea of a single global state in
Redux with some level of modularity for individual tiles (and perhaps
other UI components, too).

It takes Redux's idea of combining reducers (see link below), but
removes the need for a explicit central list of all the reducers in the
application to be maintained.  It additionally allows multiple reducers
to be assigned to the same state object through function composition,
which may or may not be a terrible idea.

Each tile would have its own reducer, taking care of managing the state
within its section of the state hierarchy.  I imagine that having the
entire app's state in a single tree (probably feasible?) would make it
super easy to offline the whole thing.

Includes the example Mail tile reworked to fit this pattern.

See also: http://rackt.org/redux/docs/api/combineReducers.html