NEWSTART-35 Implement stream access
It is now possible to ask for the n most recent items from a stream, as
well as the total number of items in the stream.

This is used within the UI to implement infinite scrolling for
notifications and to badge the notifications tab with the total number
of notifications.  (Not the unread count because we don't have that
concept on the front-end at all, yet.)