NEWSTART-46 Refuse to create activity with no recipients
Changed the endpoints to
POST /api/streams/:appId/[notifications|activities], accepting a single
item.

If the activity is created successfully, its ID is returned.  Otherwise,
an error response states what the problem was.

Had some difficulty providing a useful error response from an action
accepting multiple activities at once.  Perhaps we can put this back in
if we work out a good way to do this.