NEWSTART-44 Push created notifications to websocket
One actor per user, in charge of receiving notifications and delivering
them to the user in whichever way is determined to be best.

A supervising 'users' actor receives messages for users and forwards
them on to individual actors, creating the actor if it does not exist.

It doesn't appear possible to arrange things so that the WebsocketActor
is created as a child of the user actor.  Instead, the WebsocketActor
registers itself with the user's actor when it is connected and
deregisters when closed - may not be bullet-proof.