NEWSTART-44 Cluster sharding of UserActor
Divide UserActor instances around the cluster.  Each holds a buffer of
messages not yet delivered to a websocket.  The buffer should be saved
for migration between nodes in the cluster - need to figure out how to
test.

WebSocket deliveries are done through pub/sub, though the WebSocketActor
lets the UserActor know when it connects/disconnects from the client.