NEWSTART-67 Clear out data when usercode changes
When the usercode of the logged-in user differs from what is recorded
locally, clear both the Redux store and local storage.

Changes the way the application reducer works so each registered reducer
is always called.  This makes it possible to return to the initial state
of the app more easily.