NEWSTART-123 Unregister service worker if signed out
If the service worker's request for new notification items is met with
an Unauthorized response, unregister the service worker so it will not
receive notifications in future.

Also, avoid displaying the default 'updated in the background'
notification by always showing some sort of notification in response to
a push.

(Also build service-worker.js through Gulp.)