NEWSTART-208 Add ID7 borders, split desktop and mobile render
The render() method was full of checks for whether it was rendering a
desktop or mobile layout.  I think it's cleaner to do this check once
and give the two (quite different) layouts separate render methods, at
the cost of a tiny bit of repetition.  Adds methods to render common
components to minimise this.