NEWSTART-144 Retry reporting errors when client is offline
The server accepts an array of errors.

Throttling calls to the postErrors function allows us to retry if the
server is unreachable, and also avoid a client spewing lots of error
messages in quick succession.

Doing a slightly strange dance with assigning to `postErrorsThrottled`
to appease eslint, as postErrors and postErrorsThrottled depend upon
each other.