NEWSTART-236 Fix tile editing glitch
Make it impossible to interrupt the editing animation by triggering
editing again, by making the long press take the same length of time as
the editing animation.

(This may seem like cheating, but interrupting the snap animation and
immediately starting a new one causes some unpleasant juddering as a
result of the timing function.  Better to just make it not happen.)