NEWSTART-254 Highlight unread notification items
When a NotificationsView is mounted, cache the notifications last-read
date, then mark all notifications as read.  Use the cached date to
highlight any notifications that are new.

If a new last-read date is fetched from the server while the view is
mounted, update the cached date.  This mechanism exists to support
people opening Start at /notifications, where the view will be mounted
before data has been fetched from the server.