NEWSTART-269 Reliably display update notice
It seems that sometimes the application cache readying an update does
not trigger the updateready event.  The applicationCache.status variable
is updated to indicate that the update is ready, but the event listener
is not called.

This is a bit silly, but reliably detects an application cache update
when Start is brought to the foreground within the iOS app.