NEWSTART-286 Make it possible to sign out
Creates a new action /ssotarget that is used as the target for
login/logout.

In the action, check whether the LTC exists. If it doesn't exist, we
must have just signed out or decided not to sign in, so delete the SSC.
We can use this logic because we know we just came from SSO so don't
have to account for LTC expiry. Then redirect back to the root.

This way, if a user signs out from Start they won't have an SSC, so we
won't incorrectly redirect them to SSO.