NEWSTART-300 Update header height immediately
Top padding is applied to the .id7-main-content-area to compensate for
the height of the fixed ID7 header.  This was only happening after the
ID7Layout component was updated.  If the component was never updated, the
correct padding would not be applied causing the top of the page to
appear shifted up from its correct position.

Fix by additionally applying the padding in componentDidMount.