NEWSTART-311 Disable dragging from tile icon
The start-drag-after-delay implementation fires a 'second' click event
when the interaction is determined to be a click/tap and not a drag.
This would dismiss the Bootstrap modal almost immediately after it was
opened.  Fix by disabling dragging from the tile icon.

Relatedly, clicking an exclamation mark to display the tooltip on a tile
that also has content.href set would open the link instead of
displaying the tooltip.  Fix that by ignoring clicks on the icon in
Tile#onClick.