NEWSTART-435 Disambiguate dates yesterday
Break news and activity date rendering into two functions - turns out
they are the same when grouped=false but now it's easier to change that
later.

Render activities and notifications in the grouped view unless they are
being displayed in the desktop popover.

Rename some stuff to clarify purpose a bit.