NEWSTART-475 Add needsFetch property to tile JSON
The news, notification and activity tiles do not need tile content to be
fetched from the server in the same way as other tiles.

I'd previously implemented this on the server by checking for a None
fetchUrl and returning a generic empty tile content response, but this
was a bit of a bodge

Now the client-side knows that it doesn't need to fetch content for
these tiles, so doesn't send a request for it at all.