NEWSTART-594 Provide a width to the grid layout
It looks like during the rotate transition, elements in the web view
receive an intermediate width value.  The grid layout is rendered at
this width value, and doesn't want to re-render after the transition
completes, so after the transition it's the wrong width (too wide, or
too narrow).

Providing a value based on the window width rather than the width of the
parent element (as WidthProvider does) fixes the problem.