Switch some modules to use ES6 style source code instead of the crappy ES6 output.
Allows the optimizer to whittle away more code when it can see the `import`s.

Added plugin to specially handle lodash, that great utility belt, so that in the output it's only importing the methods you actually use. This means not being able to use `_()` or `chain` because they pull in everything, so replaced those either with regular composition, or with the handy `flow()` function to avoid function wrapping hell.