NEWSTART-644 Just scroll the body element
This gives sensible scrolling behaviour across iOS and Android browsers
and apps.  It seems that scrolling the body is the best way to get this
to work on everything, with scroll bars, without other things scrolling
in the background.

This loses the tile zoom transition, and some performance when you press
Back because the Me view has to be re-loaded.  The latter could be fixed
by the plan to keep everything loaded but hidden.