NEWSTART-702 Fail on LESS error.
It gets emitted as an 'error' but apparently not the right kind of error for Gulp, so manually feeding it into an error callback which causes it to exit with code 1 instead of 0. Upgraded Gulp though probably not necessary.

The missing chrome.less was because the Bamboo checkout was on an older version, even though we run npm install - I ran the same command and it upgraded the package so I've no idea what's goin on there. But at least now it won't continue on with a broken build.