NEWSTART-719 React improvements.
Update most components to be pure, so they don't re-render if their
props are shallow-equal.

Break some complex props (like `user`) into just the subfields that a
component needs (like `user.authenticated`) for better shallow-equal comparison.

Fix user reducer so its state is more likely to be shallow-equal if
nothing is actually being changed, which avoids more re-renders.

Don't update HideableView if it's hidden or hiding. Avoids re-rendering
a tab when it's coming in or out of focus.

Avoid creating a unique function on TabBarItem every time.

Fix bug in AppRoot calling setState before it is mounted.

why-did-you-update still reports a lot of problems, but it's not exact
and can report things that _might_ be a problem. I'm sure there's still
stuff to fix though.