Use more screen estate for modals plus stable height
This commit fixes NEWSTART-524 by converting the modals to use
Flexbox, with the wrapping .tile--config__modal element acting
as a container which occupies height (full page-navbar height).

Because the navbar height differs based on platform and whether the
user is using a native app or not, the HTML classes (.app, .android
etc) are added to the CSS selectors to target the individua
circumstances. For example, users in the native app do not have a
WebView based navbar. This isn't ideal, but it's difficult to better
address this without the use of platform specific code.

This has been tested in the native Android app, and in Firefox for
Android.