Update ESLint dependency
This commit updates ESLint to the latest version, along with the
AirBnB preset that was in use. As a result of this update, ~1200
new linter issues were dealt with in one way or another.

The majority of these were fixed, but some ignore rules have been
added to the .eslintrc.json file.