NEWSTART-1078 Ensure websockets are unsubscribed from PubSub.
As mentioned in the JIRA, this should just work, but there's something
screwy about our setup that means it is never informed when the WS is
closed.

All PubSub management now goes through PubSubActor, which can catch
and quietly swallow UnsubscribeAck messages that would otherwise have
been sent to a just-killed actor.

SIMPLES!!!