NEWSTART-1227 Disable Service Worker fetch listener.
Stops things from breaking for Android users while we investigate.
The app won't work offline like this, but we have an existing activity
that will tell you you're offline and take you back as soon as you are
online again. (This was added for the initial asset load but works
just as well here.)