NEWSTART-1408 Send data-only messages to FCM.
Also adjusts the message sending priority to always be HIGH, as this means it will wake up the phone to receive the message, which is what we want. data.priority is separate and used to determine the prominence of the notification on pre-Oreo devices, and this is unchanged.

We would only want regular-priority delivery if it was genuinely a data-only operation, like some background info sync or something like that (which we could implement in future if we felt like it)